task-08: Make `foo_data` statically allocated
Change `foo_data` from a void pointer with dynamically allocated space
to a char array with statically allocated space.
This is better in case of this simple task, as it is specified that data
must be at most 1 page long. Also, vmalloc was not the best choice for
kernel code.

Signed-off-by: Patrick Barsanti <patrick.barsanti@amarulasolutions.com>